feat(LinearAlgebra/PiTensorProduct): add version of `subsingletonEquiv` for dependent case
Dependent `PiTensorProduct`s over singleton types occur naturally when
using `tmulEquivDep` to split off one index from a dependent
`PiTensorProduct`. This PR creates `subsingletonEquivDep` to deal with
this situation.

The non-dependent `subsingletonEquiv` is re-defined as a specialization
of the general case. This breaks `toDirectSum_ι` in
TensorAlgebra/ToTensorPower, which relies on the precise definition of
`subsingletonEquiv`. Its proof is therefore changed to use `simp`
lemmas.